Gather observations and responses in config
