+++ src/ert/config/__init__.py
@@ -1,6 +1,5 @@
 from .analysis_config import AnalysisConfig
 from .analysis_module import AnalysisModule, ESSettings, IESSettings
-from .enkf_observation_implementation_type import EnkfObservationImplementationType
 from .ensemble_config import EnsembleConfig
 from .ert_config import ErtConfig
 from .ert_plugin import CancelPluginException, ErtPlugin
@@ -9,11 +8,9 @@ from .ext_param_config import ExtParamConfig
 from .external_ert_script import ExternalErtScript
 from .field import Field, field_transform
 from .forward_model import ForwardModel
-from .gen_data_config import GenDataConfig
 from .gen_kw_config import GenKwConfig, PriorDict, TransferFunction
 from .lint_file import lint_file
 from .model_config import ModelConfig
-from .observations import EnkfObs
 from .parameter_config import ParameterConfig
 from .parsing import (
     AnalysisMode,
@@ -30,9 +27,11 @@ from .queue_config import (
     queue_positive_number_options,
     queue_string_options,
 )
-from .response_config import ResponseConfig
-from .summary_config import SummaryConfig
-from .summary_observation import SummaryObservation
+from .responses.gen_data_config import GenDataConfig
+from .responses.response_config import ResponseConfig
+from .responses.response_properties import ResponseTypes
+from .responses.summary_config import SummaryConfig
+from .responses.summary_observation import SummaryObservation
 from .surface_config import SurfaceConfig
 from .workflow import Workflow
 from .workflow_job import WorkflowJob
@@ -45,8 +44,6 @@ __all__ = [
     "ConfigValidationError",
     "ConfigValidationError",
     "ConfigWarning",
-    "EnkfObs",
-    "EnkfObservationImplementationType",
     "EnsembleConfig",
     "ErtConfig",
     "ErtPlugin",
@@ -66,6 +63,7 @@ __all__ = [
     "QueueConfig",
     "QueueSystem",
     "ResponseConfig",
+    "ResponseTypes",
     "SummaryConfig",
     "SummaryObservation",
     "SurfaceConfig",
