+++ src/ert/config/commons.py
@@ -1,0 +1,43 @@
+from dataclasses import dataclass
+from datetime import datetime
+from pathlib import Path
+from typing import Any, List, Sequence, Tuple
+
+import numpy as np
+import numpy.typing as npt
+
+
+class CustomDict(dict):  # type: ignore
+    """Used for converting types that can not be serialized
+    directly to json
+    """
+
+    def __init__(self, data: List[Tuple[Any, Any]]) -> None:
+        for i, (key, value) in enumerate(data):
+            if isinstance(value, Path):
+                data[i] = (key, str(value))
+            if isinstance(value, set):
+                data[i] = (key, list(value))
+        super().__init__(data)
+
+
+@dataclass(eq=False)
+class Refcase:
+    start_date: datetime
+    keys: List[str]
+    dates: Sequence[datetime]
+    values: npt.NDArray[Any]
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, Refcase):
+            return False
+        return bool(
+            self.start_date == other.start_date
+            and self.keys == other.keys
+            and self.dates == other.dates
+            and np.all(self.values == other.values)
+        )
+
+    @property
+    def all_dates(self) -> List[datetime]:
+        return [self.start_date] + list(self.dates)
