--- src/ert/config/enkf_observation_implementation_type.py
@@ -1,6 +1,0 @@
-from enum import Enum, auto
-
-
-class EnkfObservationImplementationType(Enum):
-    GEN_OBS = auto()
-    SUMMARY_OBS = auto()
