+++ src/ert/config/ensemble_config.py
@@ -3,33 +3,30 @@ from __future__ import annotations
 import logging
 import os
 from collections import Counter
-from dataclasses import dataclass
-from datetime import datetime
 from typing import (
     Any,
     Dict,
     List,
     Optional,
-    Sequence,
     Type,
     Union,
     no_type_check,
     overload,
 )
 
-import numpy as np
-import numpy.typing as npt
+import xarray as xr
 
+from ert.config.responses.response_config import ResponseConfig
 from ert.field_utils import get_shape
 
 from ._read_summary import read_summary
+from .commons import Refcase
 from .field import Field
-from .gen_data_config import GenDataConfig
 from .gen_kw_config import GenKwConfig
 from .parameter_config import ParameterConfig
 from .parsing import ConfigDict, ConfigKeys, ConfigValidationError
-from .response_config import ResponseConfig
-from .summary_config import SummaryConfig
+from .responses.gen_data_config import GenDataConfig
+from .responses.summary_config import SummaryConfig
 from .surface_config import SurfaceConfig
 
 logger = logging.getLogger(__name__)
@@ -49,28 +46,6 @@ def _get_abs_path(file: Optional[str]) -> Optional[str]:
     if file is not None:
         file = os.path.realpath(file)
     return file
-
-
-@dataclass(eq=False)
-class Refcase:
-    start_date: datetime
-    keys: List[str]
-    dates: Sequence[datetime]
-    values: npt.NDArray[Any]
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, Refcase):
-            return False
-        return bool(
-            self.start_date == other.start_date
-            and self.keys == other.keys
-            and self.dates == other.dates
-            and np.all(self.values == other.values)
-        )
-
-    @property
-    def all_dates(self) -> List[datetime]:
-        return [self.start_date] + list(self.dates)
 
 
 class EnsembleConfig:
@@ -94,6 +69,7 @@ class EnsembleConfig:
         self._grid_file = _get_abs_path(grid_file)
         self.parameter_configs: Dict[str, ParameterConfig] = {}
         self.response_configs: Dict[str, ResponseConfig] = {}
+        self.observations: Dict[str, xr.Dataset] = {}
         self.refcase = refcase
         self.eclbase = eclbase
