+++ src/ert/config/responses/observation_vector.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Dict, Iterable, Union
 
-from .enkf_observation_implementation_type import EnkfObservationImplementationType
 from .general_observation import GenObservation
+from .response_properties import ResponseTypes
 from .summary_observation import SummaryObservation
 
 if TYPE_CHECKING:
@@ -13,9 +13,9 @@ if TYPE_CHECKING:
 
 @dataclass
 class ObsVector:
-    observation_type: EnkfObservationImplementationType
-    observation_key: str
-    data_key: str
+    observation_type: ResponseTypes
+    observation_name: str
+    response_name: str
     observations: Dict[Union[int, datetime], Union[GenObservation, SummaryObservation]]
 
     def __iter__(self) -> Iterable[Union[SummaryObservation, GenObservation]]:
