+++ src/ert/config/responses/summary_observation.py
