+++ src/ert/libres_facade.py
@@ -21,16 +21,17 @@ from pandas import DataFrame
 from ert.analysis import AnalysisEvent, SmootherSnapshot, smoother_update
 from ert.callbacks import forward_model_ok
 from ert.config import (
-    EnkfObservationImplementationType,
     ErtConfig,
     Field,
     GenKwConfig,
+    ResponseTypes,
 )
 from ert.data import MeasuredData
 from ert.data._measured_data import ObservationError, ResponseError
 from ert.load_status import LoadResult, LoadStatus
 from ert.storage import Ensemble
 
+from .config.responses.observations import EnkfObs
 from .enkf_main import EnKFMain, ensemble_context
 
 _logger = logging.getLogger(__name__)
@@ -39,7 +40,6 @@ if TYPE_CHECKING:
     import numpy.typing as npt
 
     from ert.config import (
-        EnkfObs,
         PriorDict,
         WorkflowJob,
     )
@@ -189,14 +189,15 @@ class LibresFacade:
         return loaded
 
     def get_observations(self) -> "EnkfObs":
-        return self.config.enkf_obs
+        return self.config.observations
 
     def get_data_key_for_obs_key(self, observation_key: str) -> str:
-        obs = self.config.enkf_obs[observation_key]
-        if obs.observation_type == EnkfObservationImplementationType.SUMMARY_OBS:
-            return list(obs.observations.values())[0].summary_key  # type: ignore
+        obs = self.config.observations[observation_key]
+        response_type = obs.attrs["response"]
+        if response_type == ResponseTypes.SUMMARY:
+            return list(obs.observations.values())[0].summary_key
         else:
-            return obs.data_key
+            return obs.response_name
 
     @staticmethod
     def load_all_misfit_data(ensemble: Ensemble) -> DataFrame:
