+++ tests/unit_tests/config/test_general_observation.py
@@ -1,29 +1,29 @@
 import numpy as np
 import pytest
 
-from ert.config import EnkfObs
-from ert.config.general_observation import GenObservation
+from ert.config import GenDataConfig
+from ert.config.responses.general_observation import GenObservation
 
 
 def test_create(tmp_path):
     with pytest.raises(ValueError):
-        gen_obs = EnkfObs._create_gen_obs()
+        gen_obs = GenDataConfig._create_gen_obs()
 
     with open(tmp_path / "obs1.txt", "w", encoding="utf-8") as f:
         f.write("10  5  12 6\n")
 
     with pytest.raises(ValueError):
-        gen_obs = EnkfObs._create_gen_obs(
+        gen_obs = GenDataConfig._create_gen_obs(
             scalar_value=(1, 2), obs_file=str(tmp_path / "obs1.txt")
         )
 
     with pytest.raises(TypeError):
-        gen_obs = EnkfObs._create_gen_obs(scalar_value=1)
+        gen_obs = GenDataConfig._create_gen_obs(scalar_value=1)
 
     with pytest.raises(IOError):
-        gen_obs = EnkfObs._create_gen_obs(obs_file="does/not/exist")
+        gen_obs = GenDataConfig._create_gen_obs(obs_file="does/not/exist")
 
-    gen_obs = EnkfObs._create_gen_obs(
+    gen_obs = GenDataConfig._create_gen_obs(
         obs_file=str(tmp_path / "obs1.txt"), data_index="10,20"
     )
     assert gen_obs == GenObservation(
